Changed folder structure, now it's quite simple, remoived .idea, this folder is meant to be per environment, so it should not be in the branch. Made buttons to switch between graph/table. Made footer instead of the links on the bottom.
