+++ assets/bjerrum_plot.csv
