+++ assets/datenschutz.html
@@ -31,6 +31,22 @@
         .breadcrumb-navigation {
             margin-bottom: 20px;
         }
+        footer {
+            left: 0;
+            bottom: 0;
+            width: 100%;
+            background-color: #333;
+            color: white;
+            text-align: center;
+            padding: 10px 0;
+        }
+
+        footer a {
+            font-size: 16px;
+            margin: 0 5px;
+            color: white;
+            text-decoration: none;
+        }
     </style>
 </head>
 
@@ -139,5 +155,9 @@
         </div>
     </div>
 </body>
-
+<footer>
+    <a href="/assets/imprint.html">Impressum</a>
+    <a href="/assets/datenschutz.html">Datenschutz</a>
+    <a href="/assets/barrierefreiheitserklaerung.html">Barrierefreiheit</a>
+</footer>
 </html>
