--- pages/assets/page.html
@@ -1,14 +1,0 @@
-<html lang="en">
-  <head>
-    <script src="jquery.js"></script>
-    <script>
-    $(function(){
-      $("#includedContent").load("/assets/page.html");
-    });
-    </script>
-  </head>
-
-  <body>
-     <div id="includedContent"></div>
-  </body>
-</html>
