--- templates/about.html
@@ -1,7 +1,0 @@
-<!DOCTYPE html>
-<html>
-<body>
-<h1>About me</h1>
-<p>This is a portfolio site about anything that can be put in a portfolio.</p>
-</body>
-</html>
