--- templates/home.html
@@ -1,7 +1,0 @@
-<!DOCTYPE html>
-<html>
-<body>
-<h1>My Personal Website</h1>
-<p>Hi, this is my personal website.</p>
-</body>
-</html>
